Layout and buttons on iPad
